requirements.txt
Flask==1.1.2
Flask-Env==2.0.0
Flask-Negotiate==0.1.0
Flask-Script==2.0.6
requests==2.23.0
inject==4.1.2
pytests==5.4.1
